refactor(backend): isolate runtime installation probing
